db_backup.py
Monthly dumpfile retention WIP

+++ nightly_db_backup/db_backup.py
@@ -5,7 +5,7 @@
 :file:     db_backup.py
 :language: python3
 :author:   Peter Bailie (Systems Programmer, Dept. of Computer Science, RPI)
-:date:     August 22 2018
+:date:     August 28 2018
 
 This script will take backup dumps of each individual Submitty course
 database.  This should be set up by a sysadmin to be run on the Submitty
@@ -30,6 +30,7 @@ university's Submitty database and file system.
 """
 
 import argparse
+import calendar
 import datetime
 import json
 import os
@@ -43,13 +44,15 @@ DB_CONFIG_PATH = '/usr/local/submitty/config/database.json'
 # WHERE DUMP FILES ARE WRITTEN
 DUMP_PATH = '/var/local/submitty/submitty-dumps'
 
-def delete_obsolete_dumps(working_path, expiration_stamp):
+def delete_obsolete_dumps(working_path, monthly_retention_stamp, expiration_stamp):
 	"""
 	Recurse through folders/files and delete any obsolete dump files
 
 	:param working_path:     path to recurse through
+	:param monthly_retention day of month that dump is always preserved (val < "01" when disabled)
 	:param expiration_stamp: date to begin purging old dump files
 	:type working_path:      string
+	:type monthly_retention  string
 	:type expiration_stamp:  string
 	"""
 
@@ -62,7 +65,7 @@ def delete_obsolete_dumps(working_path, expiration_stamp):
 	for file in files_list:
 		if os.path.isdir(file):
 			# If the file is a folder, recurse
-			delete_obsolete_dumps(file, expiration_stamp)
+			delete_obsolete_dumps(file, monthly_retention_stamp, expiration_stamp)
 		else:
 			# File date was concat'ed into the file's name.  Use regex to isolate date from full path.
 			# e.g. "/var/local/submitty-dumps/s18/cs1000/180424_s18_cs1000.dbdump"
@@ -77,7 +80,7 @@ def delete_obsolete_dumps(working_path, expiration_stamp):
 			match = re.search('(\d{6}_\w+_\w+\.dbdump)$', file)
 			if match is not None:
 				file_date_stamp = match.group(0)[0:6]
-				if file_date_stamp <= expiration_stamp:
+				if file_date_stamp <= expiration_stamp and monthly_retention_stamp != file_date_stamp[4:6]:
 					os.remove(file)
 
 def main():
@@ -89,11 +92,13 @@ def main():
 
 	# READ COMMAND LINE ARGUMENTS
 	# Note that -t and -g are different args and mutually exclusive
-	parser = argparse.ArgumentParser(description='Dump all Submitty databases for a particular academic term.')
-	parser.add_argument('-e', action='store', nargs='?', type=int, default=0, help='Set number of days expiration of older dumps (default: no expiration).', metavar='days')
+	parser = argparse.ArgumentParser(description='Dump all Submitty databases for a particular academic term.', prefix_chars='-', add_help=True)
+	parser.add_argument('-e', action='store', type=int, default=0, help='Set number of days expiration of older dumps (default: no expiration).', metavar='days')
+	parser.add_argument('-m', action='store', type=int, default=0, choices=range(0,29), help='Day of month to ALWAYS retain a dumpfile (default: no monthly retention).', metavar='day of month')
 	group = parser.add_mutually_exclusive_group(required=True)
-	group.add_argument('-t', action='store', nargs='?', type=str, help='Set the term code.', metavar='term code')
+	group.add_argument('-t', action='store', type=str, help='Set the term code.', metavar='term code')
 	group.add_argument('-g', action='store_true', help='Guess term code based on calender month and year.')
+
 	args = parser.parse_args()
 
 	# Get current date -- needed throughout the script, but also used when guessing default term code.
@@ -111,6 +116,9 @@ def main():
 		semester = 's' + year if today.month <= 5 else ('f' + year if today.month >= 8 else 'u' + year)
 	else:
 		semester = args.t
+
+	# MONTHLY RETENTION DATE
+	monthly_retention_stamp = "{:0>2}".format(args.m)
 
 	# GET DATABASE CONFIG FROM SUBMITTY
 	fh = open(DB_CONFIG_PATH, "r")
@@ -175,7 +183,7 @@ def main():
 		working_path = "{}/{}".format(DUMP_PATH, semester)
 
 		# RECURSIVELY CULL OBSOLETE DUMPS
-		delete_obsolete_dumps(working_path, expiration_stamp)
+		delete_obsolete_dumps(working_path, monthly_retention_stamp, expiration_stamp)
 
 if __name__ == "__main__":
 	main()
